feat: autenticação por token.

+++ app/routes/menu_routes.py
@@ -1,8 +1,10 @@
-from datetime import datetime
+from datetime import date
 from typing import Any
 
 from flask_smorest import Blueprint
+from flask_smorest import abort
 
+from app.auth import is_authenticated
 from app.extensions import db
 from app.models import Menu
 from app.schemas import MenuInSchema, MenuPutSchema, MenuOutSchema
@@ -27,8 +29,13 @@ def list_menus():
 @menu_blp.arguments(MenuInSchema)
 @menu_blp.response(201, MenuOutSchema)
 def create_menu(menu: Menu):
-    # TODO: Checar se já tem um menu cadastrado na data enviada (menu.date)
-    # Deve ser feita uma query e se o resultado não for nulo, jogue 409 conflict
+    if not is_authenticated():
+        abort(401, message="O token não foi enviando ou é inválido")
+
+    stmt = db.select(Menu).where(Menu.date == menu.date).limit(1)
+    if db.session.scalars(stmt).first():
+        abort(409, message="Já existe um menu cadastrado para esta data")
+
     db.session.add(menu)
     db.session.commit()
     return menu
@@ -39,8 +46,13 @@ def create_menu(menu: Menu):
 @menu_blp.response(200, MenuOutSchema)
 @menu_blp.alt_response(404)
 def update_menu(menu_data: dict[str, Any], menu_id: int):
-    # TODO: Checar se já tem um menu cadastrado na data enviada atualizada (menu.date)
-    # Deve ser feita uma query e se o resultado não for nulo, jogue 409 conflict
+    if not is_authenticated():
+        abort(401, message="O token não foi enviando ou é inválido")
+
+    stmt = db.select(Menu).where(Menu.date == menu_data["date"], Menu.id != menu_id).limit(1)
+    if db.session.scalars(stmt).first():
+        abort(409, message="Já existe um menu cadastrado para esta data")
+
     menu = db.get_or_404(Menu, menu_id)
     for key, value in menu_data.items():
         setattr(menu, key, value)
@@ -52,15 +64,22 @@ def update_menu(menu_data: dict[str, Any], menu_id: int):
 @menu_blp.route("/<int:menu_id>", methods=["DELETE"])
 @menu_blp.response(204)
 def delete_menu(menu_id: int):
+    if not is_authenticated():
+        abort(401, message="O token não foi enviando ou é inválido")
+
     menu = db.get_or_404(Menu, menu_id)
     db.session.delete(menu)
     db.session.commit()
 
 
-@menu_blp.route("/now", methods=["GET"])
+@menu_blp.route("/today", methods=["GET"])
 @menu_blp.response(200, MenuOutSchema)
 @menu_blp.alt_response(404)
 def get_today_menu():
-    # TODO: Deve fazer uma query para um cardápio cadastrado na data atual e retorna 200 OK
-    # Caso contrario deve retornar 404 NOT FOUND
-    pass
+    today = date.today()
+    stmt = db.select(Menu).where(Menu.date == today).limit(1)
+    menu = db.session.scalars(stmt).first()
+    if menu is None:
+        abort(404, message="Nenhum cardápio encontrado para hoje")
+
+    return menu
